+++ apps/mvp/app/(website)/page.tsx
@@ -12,6 +12,7 @@ const {sanityFetch, SanityLive} = defineLive({
   client,
   serverToken: token,
   browserToken: token,
+  strict: true,
 })
 
 async function CachedIndexPage({
@@ -21,6 +22,7 @@ async function CachedIndexPage({
   perspective: LivePerspective
   stega: boolean
 }) {
+  'use cache'
   const {data, tags} = await sanityFetch({
     query: postsQuery.query,
     perspective,
@@ -97,7 +99,7 @@ export default async function IndexPage() {
           Open Studio
         </Link>
       </div>
-      <SanityLive />
+      <SanityLive includeDrafts={isDraftMode} />
     </>
   )
 }