+++ apps/mvp/next.config.ts
@@ -1,9 +1,13 @@
 import type {NextConfig} from 'next'
+import {sanity} from 'next-sanity/live/cache-life'
 
 const nextConfig: NextConfig = {
   // basePath: process.env.NEXT_PUBLIC_TEST_BASE_PATH,
   // trailingSlash: true,
+  cacheComponents: true,
+  cacheLife: {default: sanity},
   experimental: {
+    cachedNavigations: true,
     instantNavigationDevToolsToggle: true,
     instantInsights: {validationLevel: 'warning'},
   },